Show real monument details in shared transition
